Removed cascade

+++ PP_2_2_1_spring_hibernate/2.2.1. spring_hibernate/src/main/java/hiber/dao/UserDaoImp.java
@@ -27,7 +27,7 @@ public class UserDaoImp implements UserDao {
    }
 
    @Override
-   public User getUserByCar(int car_id, int series) {
+   public User getUserByCar(Long car_id, int series) {
       return (User) sessionFactory.getCurrentSession().createQuery(
               "from User user " +
                       "where user.car.id = :car_id " +
