+++ PP_2_2_1_spring_hibernate/2.2.1. spring_hibernate/src/main/java/hiber/service/UserServiceImp.java
@@ -28,7 +28,7 @@ public class UserServiceImp implements UserService {
 
    @Transactional
    @Override
-   public User getUserByCar(int car_id, int series) {
+   public User getUserByCar(Long car_id, int series) {
       return userDao.getUserByCar(car_id, series);
    }
 }
